separated container_utils into class model
got rid of separation of code in job_worker
